feat: refactored functions into discrete modules, retaining stable API surface
